[TASK] Restructure the BE module controllers
Fixes #2632
