+++ Classes/Controller/BackEnd/EventController.php
@@ -6,7 +6,6 @@ namespace OliverKlee\Seminars\Controller\BackEnd;
 
 use OliverKlee\Seminars\Csv\CsvDownloader;
 use OliverKlee\Seminars\Csv\CsvResponse;
-use OliverKlee\Seminars\Domain\Repository\Event\EventRepository;
 use Psr\Http\Message\ResponseInterface;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
@@ -16,10 +15,6 @@ use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
  */
 class EventController extends ActionController
 {
-    use EventStatisticsTrait;
-    use PageUidTrait;
-    use PermissionsTrait;
-
     /**
      * @var non-empty-string
      */
@@ -29,31 +24,6 @@ class EventController extends ActionController
      * @var non-empty-string
      */
     private const TABLE_NAME = 'tx_seminars_seminars';
-
-    /**
-     * @var EventRepository
-     */
-    private $eventRepository;
-
-    public function injectEventRepository(EventRepository $repository): void
-    {
-        $this->eventRepository = $repository;
-    }
-
-    public function indexAction(): void
-    {
-        $pageUid = $this->getPageUid();
-
-        $this->view->assign('permissions', $this->permissions);
-        $this->view->assign('pageUid', $pageUid);
-
-        $events = $this->eventRepository->findByPageUidInBackEndMode($pageUid);
-        $this->eventRepository->enrichWithRawData($events);
-        foreach ($events as $event) {
-            $this->eventStatisticsCalculator->enrichWithStatistics($event);
-        }
-        $this->view->assign('events', $events);
-    }
 
     /**
      * @param 0|positive-int $pageUid
@@ -75,6 +45,7 @@ class EventController extends ActionController
 
             return $csvContent;
         }
+
         // 11LTS path
         return GeneralUtility::makeInstance(CsvResponse::class, $csvContent, self::CSV_FILENAME);
     }
