+++ ext_tables.php
@@ -11,7 +11,8 @@ $boot = static function (): void {
         'events',
         '',
         [
-            \OliverKlee\Seminars\Controller\BackEnd\EventController::class => 'index, exportCsv',
+            \OliverKlee\Seminars\Controller\BackEnd\ModuleController::class => 'overview',
+            \OliverKlee\Seminars\Controller\BackEnd\EventController::class => 'exportCsv',
             \OliverKlee\Seminars\Controller\BackEnd\RegistrationController::class
             => 'showForEvent, exportCsvForEvent',
             \OliverKlee\Seminars\Controller\BackEnd\EmailController::class => 'compose, send',
